src/simulation/requirements.txt
aio-pika==9.5.4
aiormq==6.8.1
annotated-types==0.7.0
asttokens==3.0.0
certifi==2024.12.14
charset-normalizer==3.4.1
cloudpickle==3.1.1
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.2.0
Farama-Notifications==0.0.4
filelock==3.17.0
fonttools==4.55.8
gymnasium==1.0.0
idna==3.10
ipython==8.31.0
jedi==0.19.2
kiwisolver==1.4.8
lxml==5.3.0
matplotlib==3.10.0
matplotlib-inline==0.1.7
mediapy==1.2.2
multidict==6.1.0
numpy==2.2.2
opencv-python==4.11.0.86
packaging==24.2
pamqp==3.3.0
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
progressbar==2.5
prompt_toolkit==3.0.50
requests
propcache==0.2.1
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.6
pydantic_core==2.27.2
pygame==2.6.1
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
scipy==1.15.1
seaborn==0.13.2
shapely==2.0.6
six==1.17.0
stack-data==0.6.3
tomli==2.2.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
yapf==0.43.0
yarl==1.18.3
